Create README.md
Hi, I'm WolfnoMan, I'm fairly new to grithub but have been involved in crypto specifically for over 3 years now, my main focus now is learning the way the blockchain works and how i can better myself with my trading and basic knowledge of cryptocurrency. I've invested a lot of time into this market and do believe that their is a lot of potential in the digital currency wave.